+++ sdk/instrumentation/instrumentation-opentelemetry/src/instrumenter.ts
@@ -4,13 +4,13 @@
 import {
   Instrumenter,
   InstrumenterSpanOptions,
-  SpanStatus,
   TracingContext,
   TracingSpan,
   TracingSpanContext
 } from "@azure/core-tracing";
 
-import { trace, context, Span, SpanStatusCode, SpanAttributeValue } from "@opentelemetry/api";
+import { trace, context } from "@opentelemetry/api";
+import { OpenTelemetrySpanWrapper } from "./spanWrapper";
 
 import {
   toTracestateHeader,
@@ -73,54 +73,5 @@ export class OpenTelemetryInstrumenter implements Instrumenter {
     }
 
     return headers;
-  }
-}
-
-export class OpenTelemetrySpanWrapper implements TracingSpan {
-  private _span: Span;
-
-  constructor(span: Span) {
-    this._span = span;
-  }
-
-  setStatus(status: SpanStatus): void {
-    if (status.status === "error") {
-      if (status.error) {
-        this._span.setStatus({ code: SpanStatusCode.ERROR, message: status.error.toString() });
-        this.recordException(status.error);
-      } else {
-        this._span.setStatus({ code: SpanStatusCode.ERROR });
-      }
-    } else if (status.status === "success") {
-      this._span.setStatus({ code: SpanStatusCode.OK });
-    }
-  }
-  setAttribute(name: string, value: unknown): void {
-    if (value !== null && value !== undefined) {
-      this._span.setAttribute(name, value as SpanAttributeValue);
-    }
-  }
-  end(): void {
-    this._span.end();
-  }
-  recordException(exception: string | Error): void {
-    this._span.recordException(exception);
-  }
-  isRecording(): boolean {
-    return this._span.isRecording();
-  }
-
-  get spanContext(): TracingSpanContext {
-    return this._span.spanContext();
-  }
-
-  /**
-   * Allows getting the wrapped span as needed.
-   * @internal
-   *
-   * @returns The underlying span
-   */
-  unwrap(): unknown {
-    return this._span;
   }
 }
